sec Multiple results: Rework storage of limits
Changing limits during a multiple measurement is more an exception than a
rule. So store limits in an extra array which ususally has one entry instead
of storing limits on each result value. This reduces size of result data and
should not be an issue for clients to extract (btw: nothing processes limits
currently - our infinite wisdom tells us to do so)

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>