zera-basemodule: replace Qt4 syntax for signal/slot by Qt5 syntax
It is not only that the new syntax is cooler because we have to be more
explicit: Syntax-highlight and compiler turn into tools avoiding us to enter

connect(this, SIGNAL(foo()), SLOT(bar()))

without any complaints. Best result is a warning spawned at runtime. Further
motivation is found at [1].

We were pointed to this by linker complaining for undefined symbols. By
removing declarations without implementation this could be fixes but we could
saw connections now pointing to nothing!!

[1] https://woboq.com/blog/new-signals-slots-syntax-in-qt5.html

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>