sourcejsonapi: remove some copy & paste by introducing getPhaseNamePrefix
Am aware switch-case has a bad reputation. But using it at one place for
returning a prefix should be ok. The advantage over if-else these days is that
in case we add another phase type the compiler would give us a warning and
warnings don't remain unnoticed since we are free of warnings.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>