sourcemodule: Introduce SourceIdKeeper
We need a way to determine if we are waiting for a transaction id and if the id
is for us.
In the old days we solved this by making zero value the knob to detect we have
no transaction pending and are not waiting. Let's do better to avoid VERY
sporadic surprises on integer wraparound.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>