Make demo device parameter work for all use cases
Use cases are:
* Production: Take device name from ModulemanagerConfig::getDevNameFromUBoot()
* Developer/Remote debug: Take device name from modulemanger_config_dev.json
  (with CMake option SET_MODMAN_DEFAULT_SESSION=modulemanger_config_dev.json)
* Demo: Take device name from -d command line param

This is yet another example for things sitting at the wrong place turning
changes into minefields: Why was the u-boot/default decision in
ModulemanagerConfig?

It was ME who put it there. So it should be me to fix it but there are too many
places created by ugly gits like me back then...

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>