sec1/sem1/spm1: Spawn warnings for power module outputs not there
Background: On COM5003 modulemanager crashed powering up int ref session. In
this case just one power module is configured for 'fo0' but sec1 is asking for
fo0,fo1,fo2.
This did not pop up earlier because starting ref session from default session
had old configurations for fo0,fo1,fo2,fo3 in pcb-server.

This is not a fix and a proper fix is difficult: We nee whole system with
pcb-server to test.

So just add warnings to support on fix.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>