All modules: Make sure module parameters are iterated with reproducible order
On writing tests for dftmodule we saw sporadic fallout and we learned that
dftmodule is the first module tested by storage hash dump with vein parameters.
To make tests reproducible use QMap instead of QHash.
Just to make clear that this does not have serious drawbacks: We changed random
order by reproducible order.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>