power1module: Make phase component descriptions more precise
Had a discussion with HuHü: It showed that is not easy to understand where to
get power values in EMOB DC sessions from.
The problem is that initially the channel / index link was fix 1/2/3/SUM.
Unfortunately we have shipped and cannot change SCPI structure with current
design (the-cat-is-out-problem).

So just change the SCPI doc to explain better what the values are.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>