cro_channel: Add a signal notifying that range change was reported
On trying to use cro_channel we learned that current design misses important:
What if a user asks for channel range information while channel ranges are not
fetched completely?
Not that a complete solution is on the table but it seems clear that the least
thing to do is to make users aware: They could now deduce times better not to
ask for range details: After sigRangeChangeReported and before
sigFetchComplete.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>