Revert "adjustmentmodule: Allow proper gain adjustment of inverted channels"
We missed important: Corrections on gain are calculated based on RMS which is
always positive. So in our tests we tested impossible.
Let's go back and introduce a DC gain command which calculates GAIN based on DC
to fix this and avoid affecting AC.
Kudos to Ch. Müller suggesting this.

This reverts commit 5518c6ed82130a3ba8901dd6cd75616d01e120ad.