Start work on new value injection implementation
The test implementation has one major design flaw:
Knowing the sequence of modules is mandatory to not see crashers or masty
effects.

It has been pain already at least once: The initial implementation done in
cAdjustmentModule did not match the needs of SessionExportGenerator so a new
enum dspInterfaceOrder was reverse engineered. Now again: In
test_adj_module_lem_offset_sequences we want to get as close as possible to
production and cannot accept being bound to a tailored session to make test
happy.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>